backend/requirements.txt
Flask==3.1.0
requests==2.32.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
pytest==8.3.4
mastodon.py==2.0.0